Member caching optimization
Optimize the way member cache is updated. 30-day attendance and loot factors are considered 'critical' and are updated no matter what. 90-day, lifetime and first/last raid stats are only updated when asked, and can be deferred to a cron job to run them once a day.

Rename update_all_cache to update_cache with a parameter which can be :all to update the non-critical stats described above; rename force_recache to update_cache (instance method) which takes the same parameter.