Wishlist form fixes
If a user had both "Solace of the Fallen"s on their wishlist, and went to edit the level 258 one, the item_name field wasn't set to a wow_id, so when we saved it, it was updating to the Item entry for the level 245 one. As a fix for this, we now have a hidden wow_id field which, if present in the POST, takes precedence over the item_name value. This field gets updated by the autocomplete result selection.

As a nice side effect, we no longer have the potential confusion of changing the item_name field to a wow_id when the user selects an item from the autocomplete. It's completely transparent to them now.